# book-network/src/main/java/leftsidebrain/book/book/BorrowedBookResponse.java
package leftsidebrain.book.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {
	private Integer id;
	private String title;
	private String authorName;
	private String isbn;
	private Double rate;
	private boolean returned;
	private boolean returnApproved;

}