# book-network/src/main/java/leftsidebrain/book/feedback/FeedbackResponse.java
package leftsidebrain.book.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {
	private Double note;
	private String comment;
	private boolean ownFeedBack;
}